+++ app/routes/pricing.jsx
@@ -1,5 +1,4 @@
 import { useLoaderData } from "@remix-run/react";
-import { json } from "@remix-run/cloudflare";
 import { useState } from 'react'
 
 import FeatureTable from '../components/pricing/featureTable.jsx'
@@ -7,23 +6,25 @@ import Footer from "../components/footer";
 import Navbar from "../components/navbar/navbar";
 import PageHeading from "../components/pageHeading.jsx";
 import PricingSection from '../components/pricing/pricingSection.jsx'
-import * as pricing from '../resources/pricing.json'
-import * as features from '../resources/features.json'
+import * as pricing from '../data/pricing.json'
+import * as features from '../data/features.json'
+import * as app from '../data/app.json'
 
 export const loader = async () => {
     const pricingData = pricing
     const featureData = features
+    const appData = app
 
-    return json({ pricingData: pricingData, featureData: featureData });
+    return Response.json({ appData: appData, pricingData: pricingData, featureData: featureData });
 };
 
 export default function Pricing() {
-    const { pricingData, featureData } = useLoaderData();
+    const { appData, pricingData, featureData } = useLoaderData();
     const [enabled, setEnabled] = useState(false)
 
     return (
         <>
-            <Navbar />
+            <Navbar navBarLinks = {appData.navBarLinks} />
             <PageHeading heading={pricingData.pricingHeading} subheading={pricingData.pricingDescription} />
             <PricingSection plans={pricingData.plans}
                 pricingHeading=""